+++ ext_tables.php
@@ -1,8 +1,24 @@
 <?php
 defined('TYPO3_MODE') or die('Access denied.');
 
-call_user_func(function ($namespace, $extkey) {
+call_user_func(function($namespace, $extkey) {
   $extensionSignature = \TYPO3\CMS\Core\Utility\GeneralUtility::underscoredToUpperCamelCase($namespace . '.' . $extkey);
+
+  // === Icons ===
+
+  $iconRegistry = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(
+    \TYPO3\CMS\Core\Imaging\IconRegistry::class
+  );
+
+  // --- Spacer Content Element ---
+
+  $iconIdentifier = 'spacer_content_element';
+
+  $iconRegistry->registerIcon(
+    "{$extkey}-{$iconIdentifier}",
+    \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+    ['source' => "EXT:{$extkey}/Resources/Public/Icons/ContentElements/SpacerContentElement.svg"]
+  );
 
   // === TypoScript ===